Addressing suggestions
