+++ crates/common_lang_types/Cargo.toml
@@ -8,3 +8,4 @@ license = "MIT"
 [dependencies]
 intern = { git = "https://github.com/facebook/relay.git", tag = "v16.2.0" }
 string_key_newtype = { path = "../string_key_newtype" }
+serde = "1.0.197"
