+++ crates/isograph_schema/src/magic_mutation_fields.rs
@@ -1,9 +1,9 @@
 use common_lang_types::{
     DirectiveArgumentName, DirectiveName, IsographObjectTypeName, Location, SelectableFieldName,
-    Span, StringLiteralValue, ValueKeyName, WithEmbeddedLocation, WithLocation, WithSpan,
+    Span, StringLiteralValue, ValueKeyName, WithLocation, WithSpan,
 };
 use graphql_lang_types::{
-    ConstantValue, GraphQLDirective, GraphQLInputValueDefinition, NameValuePair, ValueType,
+    ConstantValue, DeserializationError, GraphQLDirective, GraphQLInputValueDefinition,
 };
 use intern::{string_key::Intern, Lookup};
 use isograph_config::ConfigOptions;
@@ -11,11 +11,7 @@ use isograph_lang_types::{
     ClientFieldId, DefinedTypeId, ObjectId, ScalarFieldSelection, Selection, ServerFieldId,
     ServerFieldSelection,
 };
-use serde::de::value::SeqDeserializer;
-use serde::de::MapAccess;
-use serde::de::{self, IntoDeserializer};
 use serde::Deserialize;
-use serde::{self, Deserializer};
 
 use crate::{
     ArgumentMap, DefinedField, FieldMapItem, MutationFieldResolverActionKindInfo,
@@ -35,19 +31,9 @@ lazy_static! {
 }
 #[derive(Deserialize, Eq, PartialEq, Debug)]
 pub struct MagicMutationFieldInfo {
-    #[serde(deserialize_with = "deserialize_stringliteral")]
     path: StringLiteralValue,
     field_map: Vec<FieldMapItem>,
-    #[serde(deserialize_with = "deserialize_stringliteral")]
     field: StringLiteralValue,
-}
-
-fn deserialize_stringliteral<'de, D>(deserializer: D) -> Result<StringLiteralValue, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value = String::deserialize(deserializer)?;
-    Ok(StringLiteralValue::from(value.intern()))
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
@@ -287,7 +273,13 @@ impl UnvalidatedSchema {
         d: &GraphQLDirective<ConstantValue>,
     ) -> ProcessTypeDefinitionResult<Option<MagicMutationFieldInfo>> {
         if d.name.item == *EXPOSE_FIELD_DIRECTIVE {
-            Ok(from_graph_ql_directive(d).ok())
+            let mutation = graphql_lang_types::from_graph_ql_directive(d);
+            Ok(Some(mutation.map_err(|err| match err {
+                DeserializationError::Custom(err) => WithLocation::new(
+                    ProcessTypeDefinitionError::FailedToDeserialize(err),
+                    Location::generated(),
+                ),
+            })?))
         } else {
             Ok(None)
         }
@@ -327,80 +319,6 @@ impl UnvalidatedSchema {
     }
 }
 
-fn parse_field_map_val(
-    value: &WithLocation<ConstantValue>,
-) -> ProcessTypeDefinitionResult<Vec<FieldMapItem>> {
-    let list = match &value.item {
-        ConstantValue::List(l) => Ok(l),
-        _ => Err(WithLocation::new(
-            ProcessTypeDefinitionError::InvalidFieldMap,
-            // This is wrong, the arguments should have a span, or the whole thing should have a span
-            value.location,
-        )),
-    }?;
-
-    list.iter()
-        .map(|argument_value| {
-            let object = match &argument_value.item {
-                ConstantValue::Object(o) => Ok(o),
-                _ => Err(WithLocation::new(
-                    ProcessTypeDefinitionError::InvalidFieldMap,
-                    argument_value.location,
-                )),
-            }?;
-
-            if object.len() != 2 {
-                return Err(WithLocation::new(
-                    ProcessTypeDefinitionError::InvalidFieldMap,
-                    argument_value.location,
-                ));
-            }
-
-            let from = object
-                .iter()
-                .find(|d| d.name.item == *FROM_VALUE_KEY_NAME)
-                .ok_or_else(|| {
-                    WithLocation::new(
-                        ProcessTypeDefinitionError::InvalidFieldMap,
-                        // This is wrong, the arguments should have a span, or the whole thing should have a span
-                        argument_value.location,
-                    )
-                })?;
-
-            let from_arg = match from.value.item {
-                ConstantValue::String(s) => Ok(s),
-                _ => Err(WithLocation::new(
-                    ProcessTypeDefinitionError::InvalidFieldMap,
-                    argument_value.location,
-                )),
-            }?;
-
-            let to = object
-                .iter()
-                .find(|d| d.name.item == *TO_VALUE_KEY_NAME)
-                .ok_or_else(|| {
-                    WithLocation::new(
-                        ProcessTypeDefinitionError::InvalidFieldMap,
-                        // This is wrong, the arguments should have a span, or the whole thing should have a span
-                        argument_value.location,
-                    )
-                })?;
-
-            let to_arg = to.value.item.as_string().ok_or_else(|| {
-                WithLocation::new(
-                    ProcessTypeDefinitionError::InvalidFieldMap,
-                    argument_value.location,
-                )
-            })?;
-
-            Ok(FieldMapItem {
-                from: from_arg,
-                to: to_arg,
-            })
-        })
-        .collect::<Result<Vec<_>, _>>()
-}
-
 fn skip_arguments_contained_in_field_map(
     // TODO move this to impl Schema
     schema: &mut UnvalidatedSchema,
@@ -436,188 +354,6 @@ fn skip_arguments_contained_in_field_map(
     ))
 }
 
-#[derive(Debug)]
-struct GraphQLDirectiveDeserializer<'a> {
-    directive: &'a GraphQLDirective<ConstantValue>,
-}
-fn from_graph_ql_directive<'a, T: Deserialize<'a>>(
-    directive: &'a GraphQLDirective<ConstantValue>,
-) -> ProcessTypeDefinitionResult<T> {
-    T::deserialize(GraphQLDirectiveDeserializer { directive })
-        .map_err(|err| WithLocation::new(err, Location::generated()))
-}
-
-impl de::Error for ProcessTypeDefinitionError {
-    fn custom<T>(msg: T) -> Self
-    where
-        T: core::fmt::Display,
-    {
-        ProcessTypeDefinitionError::FailedToDeserialize(msg.to_string())
-    }
-}
-
-impl<'de> Deserializer<'de> for GraphQLDirectiveDeserializer<'de> {
-    type Error = ProcessTypeDefinitionError;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::Visitor<'de>,
-    {
-        visitor.visit_map(NameValuePairVec::new(&self.directive.arguments))
-    }
-
-    serde::forward_to_deserialize_any! {
-        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
-        bytes byte_buf option unit unit_struct newtype_struct seq tuple
-        tuple_struct map struct enum identifier ignored_any
-    }
-}
-
-struct NameValuePairVec<'a, T> {
-    arguments: &'a Vec<NameValuePair<T, ConstantValue>>,
-    field_idx: usize,
-}
-
-impl<'a, T> NameValuePairVec<'a, T> {
-    fn new(args: &'a Vec<NameValuePair<T, ConstantValue>>) -> Self {
-        NameValuePairVec {
-            arguments: args,
-            field_idx: 0,
-        }
-    }
-}
-
-impl<'de, T: ToString> MapAccess<'de> for NameValuePairVec<'de, T> {
-    type Error = ProcessTypeDefinitionError;
-
-    fn next_key_seed<K>(&mut self, seed: K) -> Result<Option<K::Value>, Self::Error>
-    where
-        K: de::DeserializeSeed<'de>,
-    {
-        if let Some(name_value_pair) = self.arguments.get(self.field_idx) {
-            return seed
-                .deserialize(NameSerializer { name_value_pair })
-                .map(Some);
-        }
-        Ok(None)
-    }
-
-    fn next_value_seed<V>(&mut self, seed: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::DeserializeSeed<'de>,
-    {
-        match self.arguments.get(self.field_idx) {
-            Some(name_value_pair) => {
-                self.field_idx += 1;
-                seed.deserialize(ValueSerializer { name_value_pair })
-            }
-            _ => Err(ProcessTypeDefinitionError::FailedToDeserialize(format!(
-                "Called deserialization of field value for a field with idx {} that doesn't exist",
-                self.field_idx
-            ))),
-        }
-    }
-}
-
-struct NameSerializer<'a, TName, TValue: ValueType> {
-    name_value_pair: &'a NameValuePair<TName, TValue>,
-}
-
-struct ValueSerializer<'a, TName, TValue: ValueType> {
-    name_value_pair: &'a NameValuePair<TName, TValue>,
-}
-
-impl<'de, TName: ToString, TValue: ValueType> Deserializer<'de>
-    for NameSerializer<'de, TName, TValue>
-{
-    type Error = ProcessTypeDefinitionError;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::Visitor<'de>,
-    {
-        visitor.visit_string(self.name_value_pair.name.item.to_string())
-    }
-
-    fn deserialize_identifier<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::Visitor<'de>,
-    {
-        visitor.visit_string(self.name_value_pair.name.item.to_string())
-    }
-
-    serde::forward_to_deserialize_any! {
-        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
-        bytes byte_buf option unit unit_struct newtype_struct seq tuple
-        tuple_struct map struct enum ignored_any
-    }
-}
-
-pub struct ConstantValueDeserializer<'de> {
-    value: &'de ConstantValue,
-}
-
-impl<'de> IntoDeserializer<'de, ProcessTypeDefinitionError> for &'de ConstantValue {
-    type Deserializer = ConstantValueDeserializer<'de>;
-
-    fn into_deserializer(self) -> Self::Deserializer {
-        ConstantValueDeserializer { value: self }
-    }
-}
-
-impl<'de> Deserializer<'de> for ConstantValueDeserializer<'de> {
-    type Error = ProcessTypeDefinitionError;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::Visitor<'de>,
-    {
-        match self.value {
-            ConstantValue::Boolean(bool) => visitor.visit_bool(*bool),
-            ConstantValue::Enum(enum_literal) => visitor.visit_string(enum_literal.to_string()),
-            ConstantValue::Float(float_value) => visitor.visit_f64(float_value.as_float()),
-            ConstantValue::Int(i_64) => visitor.visit_i64(*i_64),
-            ConstantValue::String(string) => visitor.visit_string(string.to_string()),
-            ConstantValue::Null => visitor.visit_none(),
-            ConstantValue::List(seq) => {
-                let values: Vec<&ConstantValue> = seq.iter().map(|entry| &entry.item).collect();
-                let seq_access = SeqDeserializer::new(values.into_iter());
-                visitor.visit_seq(seq_access)
-            }
-            ConstantValue::Object(obj) => {
-                let serializer = NameValuePairVec::new(&obj);
-                visitor.visit_map(serializer)
-            }
-        }
-    }
-
-    serde::forward_to_deserialize_any! {
-        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
-        bytes byte_buf option unit unit_struct newtype_struct seq tuple
-        tuple_struct map struct enum ignored_any identifier
-    }
-}
-
-impl<'de, TName> Deserializer<'de> for ValueSerializer<'de, TName, ConstantValue> {
-    type Error = ProcessTypeDefinitionError;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: de::Visitor<'de>,
-    {
-        let deserializer = ConstantValueDeserializer {
-            value: &self.name_value_pair.value.item,
-        };
-        deserializer.deserialize_any(visitor)
-    }
-
-    serde::forward_to_deserialize_any! {
-        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
-        bytes byte_buf option unit unit_struct newtype_struct seq tuple
-        tuple_struct map struct enum ignored_any identifier
-    }
-}
-
 #[cfg(test)]
 mod tests {
 
@@ -644,8 +380,7 @@ mod tests {
         Err("unexpected structure of directive".into())
     }
 
-    #[test]
-    fn test_magic_mutation_deserialization() -> Result<(), Box<dyn Error>> {
+    fn parse_mutation() -> Result<Vec<MagicMutationFieldInfo>, DeserializationError> {
         let source = "extend type Mutation
         @exposeField(
           field: \"set_pet_tagline\"
@@ -672,9 +407,16 @@ mod tests {
 
         let magic_mutations: Result<Vec<MagicMutationFieldInfo>, _> = directives
             .into_iter()
-            .map(|directive| from_graph_ql_directive::<MagicMutationFieldInfo>(&directive))
+            .map(|directive| {
+                graphql_lang_types::from_graph_ql_directive::<MagicMutationFieldInfo>(&directive)
+            })
             .collect();
-        let magic_mutations = magic_mutations?;
+        Ok(magic_mutations?)
+    }
+
+    #[test]
+    fn test_set_pet_tagline_prasing() -> Result<(), Box<dyn Error>> {
+        let magic_mutations = parse_mutation()?;
         let set_tagline_mutation = MagicMutationFieldInfo {
             path: StringLiteralValue::from("pet".intern()),
             field_map: vec![FieldMapItem {
@@ -683,6 +425,13 @@ mod tests {
             }],
             field: StringLiteralValue::from("set_pet_tagline".intern()),
         };
+        assert_eq!(magic_mutations[0], set_tagline_mutation);
+        Ok(())
+    }
+
+    #[test]
+    fn test_set_pet_bestfriend_parsing() -> Result<(), Box<dyn Error>> {
+        let magic_mutations = parse_mutation()?;
         let set_pet_best_friend = MagicMutationFieldInfo {
             path: StringLiteralValue::from("pet".intern()),
             field_map: vec![FieldMapItem {
@@ -691,7 +440,6 @@ mod tests {
             }],
             field: StringLiteralValue::from("set_pet_best_friend".intern()),
         };
-        assert_eq!(magic_mutations[0], set_tagline_mutation);
         assert_eq!(magic_mutations[1], set_pet_best_friend);
         Ok(())
     }
